+++ volc-sdk-java/src/main/java/com/volcengine/model/live/v20230101/DeleteWatermarkPresetV2Body.java
@@ -31,7 +31,7 @@ public final class DeleteWatermarkPresetV2Body  {
      * <p>:::</p>
      */
     @com.alibaba.fastjson.annotation.JSONField(name = "ID")
-    private Integer iD;
+    private Long iD;
 
     @Override
     public String toString() {
